Change the Data Type of Some Fields
Issue-ID: HOLMES-39

+++ holmes-actions/src/main/java/org/onap/holmes/common/api/entity/CorrelationRule.java
@@ -34,7 +34,7 @@ public class CorrelationRule {
     @JsonProperty
     private int enabled;
     @JsonProperty
-    private int templateID;
+    private long templateID;
     private String engineID;
     @JsonProperty
     private String engineType;
